Defined how all the components communicate with each other in order to
create a mold. Also properly implemented a mock scheduler that randomly
looks at all the available presses and randomly chooses one. The random
number generator is seeded so that our local testing can be consistent.

I implemented state for the presses. IDLE means that the press is
currently doing nothing. CREATINGCHARGE means that the extruder is
currently creating a charge for the press. RECEIVING means that the
extruder is in the process of handing off the charge to the press.
WORKING means that the press is creating the mold.